Supremely mash up the array flattener.
This patch attempts to enable the array flattener in the context of the
new ast-vtable-based method calls, like array selects/stores, and so
forth.

This could have been straight forwards, however because the array
flattener is a template, the ast objects needs to know that type so that
they can cast the smt_convt context down to the correct type. And that
means baking a type into the array ast type. Which means more templates.

This is of course crazy, and hopefully in a short while I'll have
switched all of this code to not be templated based, so we'll all be
safe again.