Alter tuple array creation api format
The 'despatch' call used to date exists to switch between constant
array of's and constant arrays. However this is a detail that the caller
shouldn't really care about. Therefore reformat the api so that tuple
array creation just takes an expression, and the 'despatching', as it
were, is an internal detail.

This might mean a bit of replicated legwork somewhere, but on the other
hand the interface is now clean-ish.