Rename smt ast type names
The current tuple and array ast methods perform actions specific to the
present tuple flattener, which (using assign semantics) flattens things
down to single variables. The current intention is to bring back the
symbol flattener as a debugging option, which requires it's own flavours
of smt ast's. Thus, they're going to need separate methods / types.

This means of course that there's going to be twice as much maintenence
because there are two tuple flatteners where there could be one. However
it is /completely impossible/ to debug the node flattener, so it is
worth every piece of effort IMO.