smt_ast shouldn't assume it can use SMT_FUNC_EQ
In the case of the tuple symbol flattener, it doesn't override assign
because it doesn't do anything funky with it, however it also doesn't
work with the SMT_FUNC_EQ function application.

Fix this by having the default assign implementation being completely
neutral, by having it call the 'eq' ast method and asserting the result.