Conform solvers to new array api
Mostly this involves redirecting symbol creation to their inbuilt symbol
creator, and forwarding the array_of stuff to the default array_of
constructor. This is all fine, because fundementally no new behaviour is
being added to these solvers.

It does mean thought that we should be able to just drop in a handle to
an array flattener and have it Just Work (TM). In reality there's going
to be some pain with bits of code not knowing what kind of symbol to
create, but that's feasible to fix (by pumping everything through
convert terminal)