Un-template the array flattener
Now it just exists as a superclass of minisat_convt, takes an smt_convt
as an argument, and does exactly what it did before. The major
difference being that all of the api calls to it go through ast methods
and through the array interface, not through a massively undefined
thingymajig.