Move array flattener selection to solve.cpp
This involves a significant remash in the process, which I've rolled
into this commit because it's difficult to separate now.

Specifically: the array solver needs to have a method call before the
formula is solved, because it has to add constraints governing array
updates and so forth. Thus, a mechanism for performing pre-solve
activity has to be added, the same is added to the tuple iface for good
measure, and everything gets lawnmowered.

While ugly, this is progress though.