Fix creation of negative integers
There are only unsigned ways of creating bitvectors in yices. This makes
sense. What we want to feed it in the case of negative integers, is the
integer value represented in twos compliment, so that all the top bits
are ones.

However, this was calling to_uint64 on BigInt, which doesn't take
account of negativity, so all the negative bits were just clipped. Which
is bad.

Fix this by calling to_int64, which does what we expect.