Fix yices array issue through assign semantics
It would seem that yices simply doesn't honour array assignments as it
should. If you feed it the following benchmark:

(set-option :produce-models true)
  (set-logic QF_AUFBV)
  (set-info :status unknown)
  (declare-fun |c::a&0#1| () (Array (_ BitVec 1) (_ BitVec 32)))
  (declare-fun |c::a&0#2| () (Array (_ BitVec 1) (_ BitVec 32)))
  (declare-fun |c::a&0#3| () (Array (_ BitVec 1) (_ BitVec 32)))
  (assert (= (_ bv0 32) (select (store (store |c::a&0#1| (_ bv0 1) (_ bv1 32)) (_ bv1 1) (_ bv1 32)) (_ bv0 1))))
  (check-sat)

Then it correctly identifies the result as unsat. But if you flatten the
intermediate stores into symbols and equality them to symbols, then it
simply doesn't perform the stores.

This is fixable by just performing ast assignments instead of
equalities, but is seriously annoying.