Replace gen_uint with saner defaults
So, what I haven't addressed up until now is that when you use things
like gen_uint, it spits out a 32 bit value. This is all fine and well
until you run in 64 bit mode, which is something we should definitely
support. As a result, I've baked a bunch of 32 bit assumptions,
everywhere.

To get around this, eliminate gen_uint (or change it's signature at
least), and replace it with one where you have to give it an int
type. Also add gen_ulong. The idea behind this is that if you're in a
position to not care about the bitwidth, it's because you know that it's
going to be a machine word. If you don't know that, then you need to
perform some analysis and actually decide on a type.