Support array equalities
There are two forms of this, unbounded and bounded arrays. For the
former, we produce equalities for each index that either array may have,
allocating an index/variable if one array doesn't have an index that the
other has. For the latter, we simply have an equality for each element.

Move the record of the elem sort to a smt_sortt seeing how we're going
to end up having non-bv array ranges shortly. Various hacking will need
to occur first though.