mk_array_symbol now takes a subtype sort arg
Seeing how the current storage for smt sorts doesn't allow linking to
the range sort, which is kind of sucky, but much more structural.
Instead, the subtype is passed down by the creator, who should have it
to hand.

The whole point of this massive patch is that in array_conv's
mk_array_symbol, we now store the actual subtypes type, rather than
storing the array type itself. Which is erronous.