Play silly buggers with array update/selects
If we have some code like this:

  struct blah;
  struct blah fgasdf[200000];
  fgasdf[0].foo = 0;
  fgasdf[0].bar = 0;

Then both a select and update to fgasdf with the same index are encoded,
because we select the value from idx 0, updated it, and write it back
in. A free value is used for the select and it's constrained later.
However when that happens, it's overwritten by the update operation
before it's constrained. Which is silly.

Work around this by identifying this corner case and encoding the
relevant assertions. There's going to have to be a large refactoring of
this flattening code at some point, not today.