Allow array creation to take an array sort
When constructing arrays, if we create an array of structs, we end up
calling the solvers mk_sort with a flattener's struct sort. Which
(should) immediately croak.

Work around this case by providing the array sort from the caller,
outside of the usual array api. This is the only place that that sort is
created, as mk_array_symbol takes a sort. All other decisions based on
whether something has come from the solver or a flattener should be
resolved through the ast methods.