Work around array read/write in a better way
When we read and write the same index at the same "time", rather than
identifying this corner case and attempting to code around it, instead
identify it in advance during select. Specifically, when we perform an
unbounded select of the same index that was last updated. In this case,
we can just return the ast of the last update (which is safe), rather
than encoding an array select.

This means we don't have to worry about it's encoding during constraint
time, and is actually an optimisation as we avoid another additional
variable in the program.