Take a first shot at unbounded array equalities
This doesn't re-instate the implementation that I've left at the bottom
of the file, however it's heavily inspired by it. It also features some
guards to ensure that the ordering of encoded array equalities doesn't
lead to unbounded array selects that are going to be missed/discarded
due to their array already having been converted.