Start supporting non-same-base array ite's.
The precondition for this, as mentioned, is that the set of array
indexes for both arrays is the same, which I've baked into the
assertions I'm filing. These are expensive, but really, really
necessary (and NDEBUG will turn them off).

Code quality here is going down because I'm in a rush. Basically, quite
a bit of this array flattener is going to need a rewrite, so I'm just
going to spray yolos everywhere.

The primary change is that we're, during constraining, making unbounded
selects of one source of the ite, and then proceeding as normal. The net
effect is performing the net effect of the ite across base array ids.

One annoyance is this means that it's unclear which array ID is going to
be the result of such an ITE: it could validly be either, because the
resulting ast fully represents the value of the array. In this case, we
pick the smaller number, because it means the other array won't have
been encoded by the time we perform this ITE, meaning we don't have to
go backwards.

This commit message probably makes no sense; too bad.