Join array indexes across different array IDs
That is; when an ITE or an equality occur between base array ID's, we
can encode it, with the precondition that the set of array indexes
between the two are identical.

Implement this by collecting the set of array IDs that each array
touches, computing a fixedpoint across them, and then joining the
resulting sets of arrays.