Fix incorrect indexing during array updates
We should be using the index stored in the index map for this, rather
than an explicit counter. This worked in the past because two maps/sets
were consistent in their ordering, but with the storage change to boost
containers that's changed.

Either way, relying on that kind of consistency, implicitly, is bad.