Fix repeated assertions in runtime encoded eqs
What happened here, was that when we pop'd a context, the counter of how
many SSA instructions had been encoded was reset to the /start/ of a
particular scope, rather than the end. That meant that whenever pushed
another context, the same set of expressions were asserted for a second
time, which interferes with some of the backends.

To fix this, keep a seperate track of how much of the current context
has been encoded, and when popping, update it to reflect the portion of
assertions in the current scope that have already been asserted.