Attempt to fix yices bitvector inequalities
If you run with --smt-symex-guard and yices on 03_nec08, we get the
wrong result and an invalid counterexample. It would appear that the
crux of the matter is that inequalities just aren't working, because if
you replace the != in the test with a < instead, everything works.

Two problems: first, we weren't using bvneq_atom. Second, is that even
that doesn't work.

Attempt to fix that by using yices_redcomp, which seems to guarentee
that it'll work on bitvectors of any size . Then compare the result with
a one-bit bitvector, using bveq_atom. My guess (this is documented
nowhere) is that bvneq only works on a single bit, rather than a whole
bv.

This is massively painful and bad, however appears to fix 03_nec08 in
incremental mode.