Enable more caching of expr CRCs
During CRCing of expressions, in the past we've just calculated it and
then discarded it. Instead, actually cache it on every single CRC that's
produced, for an expr2t.

This is still safe, because everything is contained in an expr2tc, and
you can only mutate it via calling a mutating get() method, which will
set the crc value to zero.