Don't hash outer expression data into inner exprs
The previous patch led to the seed being passed down to
expr_methods::do_crc being hashed into the expressions crc_val field.
That meant that an expressions crc depended on the context it was crc'd
in, which is bad.

Instead, begin crcing from an initial zero value. Once it's calculated
for the current expr2t and it's subfields, hash_combine it into the
passed in seed (as we _do_ want to accumulate it into that).