Fix a scoping error
In this context, where we're giving an implicit cast from the return
value of a function to the type of the callers return variable, we
declare a new code_assign2tc. However, we don't write it anywhere, and
it aliases the parameter named 'assign'.

So, explicitly actually assign it to that parameter instead.