Fix creation of array-of-arrays.
Specifically, when we have two layers of constant_array2t, we
successfully flatten it's type down to a single dimension, which is fine
when creating symbols. However when creating constant arrays, their
representation is not flattened, and we end up trying to store an array
in an array.

So, write a function that does this.