Add ltl-dummy command line option
There's a difference between what a lot of my older scripts do and what
the ltl option now does. Nowadays, --ltl causes ESBMC to solve a formula
three times, looking for BAD, FAILING or SUCCEEDING assertion hits.
Wheras in the past this was pieced together manually.

In the interests of replicating old behaviour exactly, add an option
that allows one to run single traces at a time. They have to be manually
specified through setting claims on the command line.