Remunge CXXFLAGS in automake files
It appears that AM_CXXFLAGS is supposed to be configured as CXXFLAGS for
all things built by automake. So, configure that (rather than
lib-whatever). Also don't try and SUBST in, as it needs an explicit
definition?

Finally, dump our normal CXXFLAGS into ESBMC_CXXFLAGS, and pump it into
AM_CXXFLAGS. And put in the -std=c++0x flag.