Remove automake symlinking foo
This existed because I couldn't work out how to make flex output a file
to a given filename. That turns out to be the -o option, so this
symlinking no longer needs to exist. That then requires however, the
ordering of the sources to change, so that the scanner is built before
the parser, as the parser includes the scanner.