Pick a c++ compiler, prefer clang
Add the check for whether CXX supports c++11 after picking CXX. This
triggers a bug in clang 3.0 on ubuntu, so it'll need to be over-ridden.

Ideally there'd be some kind of macro that loops through looking for the
appropriate compiler. However, I don't have the time to learn m4 enough
for that. Instead, the user will need to override with CC= and CXX= when
running configure.