Fix plain multithreaded exploration
An error occurred here -- in 88e1269 I made ESBMC always use the runtime
solver for solving things. However I forgot that this doesn't get
reallocated / cleared when we're not in smt-during-symex mode. And thus,
all exprs were converted and asserted into the same context, which would
have immediately returned unsat once it was initially found to be unsat.
So this is all particularly poor.

This probably invalidates some of the results I'm trying to present
right now, so I need to re-run a bunch of things.