Be more forgiving about failed func ptr calls
We're a bounded model checker, after all. If there's no resolution of
symbols (this seems to be to do with static, or non-function symbols?)
then print a warning

I anticipate that this is because a data symbol is being called. Which
is dumb, but happens, and should be picked up somewhere else.