Improve alignment calculation
It looks like I missed the part where, when calculating alignment after
pointer arith, the pointer alignment is of consideration, not the thing
it points at. (The actual offset into the thing is significant though).

This fix assumes that the pointer alignment is the minimum alignment
when a nondet int is added to a pointer with a known offset. If the
known offset is less than the pointer alignment (which is illegal
anyway), assume it's one byte alignment from here on.