Try to support correctly linking boolector
Pass a macro down to the solver-location-testing phase to add additional
config options to command lines, such as adding libraries and flags.
This means that we can add lingeling and minisat to the compiler command
line when testing for boolector.

That too means that we can only use this on versions of boolector from
1.6 or later, as the directory structure before that is different.
Howver I don't care, as from 2.0 onwards they've changed the API in an
incompatible way, so there's no point trying to use previous versions
anyway.