Remove const qualification from various places
These are now being reported as errors, legitimately. Possibly our
previous set of options allowed these to be compiled, but the use of
autoconf resets those options. This is fine, because there should be no
excuse for things like '-fpermissive'.

Fix them for the moment through the use of casts: this is ugly and bad,
but that's ok because the broken code is bad. Ultimately it should all
die through the use of irep2 and clang AST.