Support another flavour of GCC ASM statement
This crops up in all the TACAS benchmarks; the previous change to ASM
parsing was done on a mac, and looks like it wasn't sufficiently tested
at the time. Ugh.

I get the impression that it's a losing battle to try and support all
these asm syntaxes in the long run.