Abort if C library fails to merge
There are circumstances (such as conflicting definitions of variables)
that mean we can't link the C library and the program under test
together. Unfortunately, this was silently ignored, leading to errors
further down the line.

Fix this by calling abort whenever that happens. An error is printed;
it's up to the user to resolve it.