Remove 'time' C library implementation
This is fairly brutal, but the ESBMC built on my machine (and others)
conflicts with the headers built into some of the SV-COMP benchmarks,
and linking the two together fails (validly).

This exposes something fairly unpleasent about SV-COMP: all the files
are preprocessed, so there may be a ton of implementation specific data
in there, specific to whoever preprocessed it. In particular, we can't
just model check 'pthread' operations, because we have to parse and
understand the linux implementation of pthreads instead.

We could move back to tacking C library functions onto the end of the
program under test, however it's then difficult for the user to identify
the fact that code inside ESBMC is conflicting with their system
headers. The whole thing is, unfortunately, a trainwreck.