Start attempting to strip out scalar step list
The intention of this feature was that we took the set of "index" and
"member" expressions that converted a struct/union reference into a
scalar, and apply it in the middle of building a reference when an
appropriate type was reached. However, it turns out to be quite the
train wreck.

Better, once we make the presumption that we're going to be converting
from a struct to a scalar, is to take the explicit offset introduced by
that computation and add it early.