Correctly take address of realloc'd symbols
Just to add additional complexity to the memory address space situation,
it looks like when I originally implemented the pointer-renumbering
situation, I didn't correctly account for the normal take-address
operation. One must actually check whether a pointer has been renumbered
when creating an AST of it's address.

The result is that, with some additional changes to pointer operations,
overflow_22_wordplay fails to operate correctly, and in fact winds up
making the whole formula unsat. Which is poor.

This only happens for Z3 -- probably because by re-assigning a different
pointer object to a symbol name, there's an implicit equality, which
becomes false.