Refactor byte sticthing implementation
This separates the two tasks of actually extracting some bytes from a
data object, and the stitching of them back together into another data
object. This is important because there are several ways in which one
can do the former, but should only be one way of doing the latter.