Cope with non-byte-sized dynamic mem access
That is, previously when accessing a type within another incompatible
type (i.e. a uint16 inside a uint32), we extracted a byte, then casted
it up. Which is bad.

Instead, extract, and then stitch, a sufficient number of bytes. Might
cost more resources, but will actually be acurate for once.