Support assignment through reconstituted structs
That is, allow a constant_struct irep that has been composed from
stitched-together data during dereferencing, to be assigned to. This
ultimately decomposes into assigning to each data element the
reconstituted structure is made up of.

It's extremely nebulous to allow an assignment to what's labelled a
"constant structure", unfortunately there is no suitable irep for
representing a reconstituted one. And it would be otherwise identical.