Revert "Begin rewriting union initialization code"
This reverts commit a308c0893e3c102db400a191478711b0445b799b.

After actually thinking about this further, this transformation needs to
happen at a reasonably large number of points in conversion, for example
one can assign a struct literal that contains a union at any point in
any piece of code.

It's going to be annoying, but more technically flexible, if we rewrite
unions into byte arrays at migration time. Ugh.