Decompose union literals into byte arrays
This migration function should allow for the almost-complete elimination
of unions from the BMC process, as everything related to unions now
becomes a series of byte arrays. AFAIK, there is nothing in BMC that
actually has to consider that something is a union any more.

Note that it will be tricky to select where to apply this function: it
depends on union literals having a union type, which is going to lead to
misery if applied _after_ the union transformation process, or at the
wrong time during it.