Fix build error in eigen code
Eigen defines this to be an integer (long story, in "Matrix.h", the
final "EIGEN_MAKE_TYPEDEFS_ALL_SIZES" code). On 64 bits this
inexplicably works, but on 32 bits it's a comparison between signed and
unsigned integers.

I don't quite get why 64 bit's doesn't mind this, seeing how integers
are the same size on each platform. Only thing I can think of is that
actually the return of rows() is a long, in which case unsigned int gets
promoted to a long safely on 64 bits.