Don't rewrite result type of dereferences
We want something that dereferences a union pointer to evaluate to a
union; we can then let the dereferencing code pick this apart, rather
than trying to rewrite the surrounding expressons to operate on byte
arrays, which would become Emotional.

This was, for some reason, breaking all the pthread regression tests
that contained the --data-race-check flag.