Cope with operations on non-bit byte operations
That is; operations where byte extracts and updates are applied to
things that don't immediately have a bit representation, for example
booleans and pointers. These must be the subject of casts for them to be
moved to and from their bit representation. Handle these by wrapping in
some additional layers of casts and recursing further, so that we only
ever deal with types that have a bit representation

Of course, such casts are expensive, but there's not a lot we can do
about that.