Correctly handle some deref type mismatches
That is; when we dereference something to a field in a structure, we
ignore the types, and only care about the size. This means that we can
eventually wind up assigning an integer into a pointer, in a poorly
typed manner.

Fix this by examining the type of what we've built a reference to, and
insert a typecase as appropriate.

At the same time, handle the case where we have the correct alignment
into a struct field, but under-read it.