Weaken 'with' operation typechecking
We can have a with assigning a pointer into a pointer field, and have
the types mismatch due to the pointed-at type being slightly different.
In reality all we care about at this level is that the smt
representation matches, so just check the base type to be equivalent.

(The base type being wehther the two are compatible, in a sort-of-c like
 way).