Cope with constant arrays within array_of's
Array_of's occur when we have some kind of global array that needs
initialzing with a default value, such as zero, of a certain size.
I don't believe one can say in C "here's a multimensional array, fill
each member with this default array". One could explicitly encode that,
but not, I believe, implicitly.

The union mangling however makes that possible, by having an array of
unions, than then gets turned into an array of flattened unions (which
are themselves arrays). Boom.

Work around this by, when encountering a constant array of values within
an array of, to manually duplicate all the contents of the inner array,
then continue converting.