Expand union type writing circumstances
If we dereference to a structure reference, that contains a union, then
we definitely need that union to be flattened. This is because an
assignment built around it will need to have that array, as that's the
concrete type that these things evaluate to.

The immediate way this presents is an assertion failure when the
types of two values being assigned don't match.