Fixup integer type when performing pointer arith
It looks like there's trouble with array indexes -- they're coming out
as 32 bit integers, but everything else is using 64 bit integers when
the --64 flag is specified. To avoid this, cast to the unsigned long
type if we're not using it already.

It wasn't consistent anyway, because the other side of the branch was
using gen_ulong.