Support assignment to constant arrays
When reconstituting data structures like arrays, we need some kind of
container to store elements in: for constructing arrays out of smaller
elements (i.e. byte arrays) that'll be constant array. In the
circumstance where we assign another array to such a reconstituted one,
special handling must occur.

As with the method in symex_assign_structure, we explicitly unpack each
element of the reconstituted array and create a sub-assignment to it.
This allows for an assignment to whatever it was reconstructed from to
be encoded.

Rename the old "symex_assign_array" to "_index" to accurately reflect
what's going on: the former is an assignment of a /whole/ type of array,
while the latter is a probably-scalar assignment where the lhs happens
to feature an index.