Don't duplicate padding bytes
Previously we were adding all the additional padding bytes required on
the end of the relevant structure -- which isn't something that will
lead to a lot of accuracy. Instead, the correct action is to add that
padding at the end of relevant fields (see previous commit).

That led to adding the padding twice; fixed in this commit by not just
adding all the missing padding at the end of the struct.