Delete ancient Z3 workaround code
The 4.0 release didn't seem to have working overflow support: it would
incorrectly set the reference counts on the ASTs it produced / used. To
work around this I introduced my own implementation which created
similar ASTs. However, that's now producing bugs when run over SV-Comp.

Remove the old Z3 code; use the smt-level implementation of this. That
might be a slower way of using Z3 to solve this problem, but it actually
works.