Finally fix type methods recursion method
Kind of looks like it wouldn't match the vector correctly. The reason
why is because the pop_front 'function' doesn't produce another vector,
it produces something that can be operated on like a vector. So it can't
be matched in specialization so easily!

Rely upon boost enable/disable thingymajigs instead.