Further define recursive constructors
In the end all this does is forward from the top class down to the
bottom class, through all these layers. To protect the programmer, we'll
be relying on the top level definition (like 'symbol2t') calling with
the right arguments. That way the users don't get exposed to a million
template errors.