Nix use of decltype
As it happens, we can only get type* or type& out of decltype, because
expressions can only ever refer to it, not actually _be_ that type. We
would have to explicitly name it.

Instead, switch to naming the typedefs in the scope of the
type_methods2_rec class. For whatever reason they weren't working for me
before; it now appears that resolution works. Oh well.