Mark up when {expr,type}_id is used in methods
That is: certain of them do not consider type-id or expr-id as being
part of the structure (i.e., list operands, get operands), but do for
things like comparisons, crcing and hashing. Indicate when this happens
in comments. This is merely an ugly part of the fact that these are
considered normal fields.

At the same time, kill some base implementations of do_crc and hash,
because these can now be handled just like any other field, rather than
having to special case them. IMO, this is an improvement.