Fix some overloading in getting subexprs
For the mutable get_sub_expr, we need to produce non-const references to
the fields of the class, or crazy things will happen. But due to too
much copy and paste, we were using a const pointer to the class rather
than a mutable one, so crazy things started happening.