Use a different set of traits for not2tc
Reason being: we rely on one of the current template hacks to a) assert
that the contained type is of the right dynamic type for almost all
types, but b) force new construction in cases where it's ambiguous as to
whether the user meant a new expr, or copying one.

Specifically, not2tc(foo) might make a new not expr on top of foo, or
might upcast foo to be a not2t type. In the past we've assumed the
former. Add type traits to ensure it's configurable.