Move type traits to not just being a vector
That is; it looks like it's going to have to carry around additional
information about the type and how it's used, which should go in the
traits class. So, keep the ::type field as a vector of elements, but
change where it gets passed into the irep_methods templates. Then, the
irep methods templates can access superclass::traits::foobar to get
extra data.