Rename subclass to baseclass in irep_methods2
This was technically incorrect: a subclass is something that derives
from a superclass. Wheras this argument is a class that all these
templates derive from.

We're already using the term 'superclass', and that I guess should refer
to the /immediate/ superclass. So, use 'baseclass' instead, indicating
that it's the class at the base of our stack of templates.