Use always_construct type trait
This constructor needs different behaviour depending on the type traits:
either to always construct, or always copy-construct and assert that the
expression is correctly typed.

To do that, fetch type traits and inspect them, rather than examining
lots of template arguments. This can't actually be done by unpacking the
vector of fields as a reasonable amount of interpretation would have to
occur.