Protect address-of-const-str from poison null byte
If one embeds a null in a constant string, and then takes the address of
it, then certain model checkers (i.e. boolector) truncate the symbol
name at the null, leading to crazy behaviour later. The solution is to
not allow null bytes to go into SMT symbol names.

There's some risk here in that the address of a similar constant using a
'_' instead of null will overlap. Too bad.