Implement bidirectional pipes
We were relying on the fact that the base case is always faster
and would present the result before the forward condition or the
inductive step could (incorrectly) present the wrong result.

loops/terminator_01_false-unreach-call_false-termination.i proved
me wrong.

This commit implements bidirectional pipes, so we only present the
answer after checking if the base case finished.

Signed-off-by: Mikhail Ramalho <mikhail.ramalho@gmail.com>