Forward dec a clang class, don't bring in header
That header brings a lot of warnings with it (which knackers the esbmc
build dir) so ideally should be avoided. The only use needed is ASTUnit,
which is used as a template param to unique_ptr, thus the size isn't
needed and we can just forward dec it.

The cpp files will need to include the relevant headers themselves.