Generate a boost set of traits
This was originally a plan to explicitly instantiate all these things
via some boost iteration. But, it turns out that explicit instantiation
is miserably special cased, and we can't do it through nested name stuff
(i.e. ::).

While this is useless, I want to keep it around to keep the concepts
available to my mind.