Experience template based pain and misery
It turns out that we need an explicit instantiation to make the
python_init object get compiled into the binary -- but I've spent a lot
of time trying to make that happen via metaprogramming, but it can't be
done. This is because the instantiation syntax has to be completely
explicit or something. Highly miserable.

I've then found that this could all be done by putting the python_init
field into the derived object, which gets an explicit instantiation
anyway for other reasons. I then spent a reasonable amount of time
running into my old friend, ``couldn't deduce template argument''.

This final patch works, and should link, but represents a high volume of
misery.