Define python_init field of templates
I forgot that this isn't actually a class member -- it's a template
member. So we need to define it as a template. We can do that explicitly
later on, or we can just declare it in the header, and (apparently) c++
won't actually instantiate it until it's asked to.

Ugh.