Annotate type traits with more cons-arg info
That is, what arguments the constructor takes. This is important to be
able to determine what arguments to pass into the constructor at a later
date for the python object creation stuff.

As a result, we discover that there's a mismatch with code_skip2t, in
that it doesn't take a type, but lots of ireps like it do. So, because
nothing actually manually uses it, add a type argument.

This means that we can actually now, correctly, create
build_python_class methods that hook up to the right constructor. The
fact that this built before-ish is entirely due to the previous commit,
which I fully reckon is a compiler bug.