Turn everything into a libtool archive
This is necessary to actually be able to build shared objects like
libesbmc.so (or whatever). Libtool demands all libraries must be libtool
libraries. Sigh.

Some eratta covered by this commit: we generate .c files now instead of
pumping them through a pipe to gcc via flail.sh, because libtool won't
accept compilation of something it doesn't recognize. So make everything
appear to be a .c file.

There's also a duplicate instance of goto_program_dereference.cpp in a
makefile that I seem to have put there accidentally in the past, which
causes compilation errors, but only when linking a shared library!