Disable certain warnings across the project
Certain header inclusions on ubuntu 16.04 are exploding due to the
presence of auto_ptr in various headers, which c++11 defines as
deprecated. We'll have to wait a while for those headers to be updated
if we want to run with Werror otherwise.

(cherry picked from commit b8a0b2e0b3165e949cac9f7aed3d4faeef15f10b)