Switch python make constructor to making expr2tcs
Rather than, for example, an add2tc. The issue is that boost::python
doesn't seem to be able to discover that add2tc is a subclass of
irep_container<expr2t>, and thus convert between the two. It's probably
possible to write this or otherwise let boost::python know that this is
possible, but the extra type safety we gain is zero.

More of an issue is how we downcast from an expr reference to a
something-else reference.