Allow container-to-irep ptr conversion in python
After... about a week of work. More commenting and refinement will come,
but the general gist is that we specialize the shared_ptr_from_python
class specific to building type2t's, by just sucking the stored type2tc
out of the pointer holder, and returning the contained type2t.

Matters of pointer longevity will documented in comments in the next
commit.