Attempt support for converting None to container
Right now what we have is a good function for converting a contained
irep reference into that irep reference, so that we can access it.
We can pass it to functions that want a shared pointer too because we
just pass the container in.

However: we don't have anything that will specifically create new
containers, and so there's nothing that'll take a Python None and cough
up an empty container.

Try to support this via registering an rvalue converter that does this:
it doesn't commit to having any storage for None, so it gets handed
enough to contain a temporary conversion of None to Container. Build an
empty container in there and return it.

This works up into make_contained, after which all the pointers
simultaneously move one foot to the left. Really. This is something
that's worked up until now with no problems, so advanced debugging
techniques are required.