Have irep methods take const references
Without this, it looks like the type inferrer looks at what's being
passed into the constructor and, if it's a temporary, repeatedly
constructs additional containers around it. Which is completely
needless: we can just pass the references down the chain.

Fix this by explicitly adding const and reference to each type in the
Args template parameter pack. I think this has no effect on existing
constant references.

Performance improvement here is probably minimal, but it does fix
stepping through a bunch of things in boost.python right now.