Use boost shared pointers in various places
Basically: this is going to save a lot of effort exporting this to
boost.python, because it supports it out of the box.

This raises the questions of whether or not it would have been better to
mangle the ireps to handle this: IMO the answer is no, because a) I've
learnt sufficiently about b.p in the process, b) we'd have to maintain
the interface boost shared_ptrs expose to b.p, c) there'd be a lot of
mucking around inside b.p's namespaces.

It probably breaks even.