Export lots of symex stuff
This is pretty much every method and field in art and ex_state that we
can get our hands on, with the exception of monitor and state hashing
related stuff. Plus a few methods where we can't return an integer
reference.

Required turning on -Wno-strict-aliasing, because there's some code in a
boost header that breaks it. Ho hum.