Subclass STL container to make boost.python happy
Normally this is a terrible sin. However, unordered_map doesn't provide
any facility for comparing keys (which is fine seeing how it's a hash
map). But, boost.python expects that kind of facility for the purpose of
creating it's own proxy list of Things, with ordered indexes. Which
mildly sucks.

The only way around this without actually moving back to std::map is to
extend unordered_map, which makes me feel bad, but boost.python good.