Move trap_to_python into goto-symex
Reason being that it complicates the linking process by having things at
the top of the dependency chain (exported build-module function that
calls into objects like symex_python.o) and at the bottom of the chain
(trap_to_python). Instead put it where it's most appropriate, in the
symex facilities, because it's a symex facility.